split HomepageAutomation into GraphsAuto.,Sorting.Auto.,Queries.Auto.
