+++ src/test/java/com/generic/functional/automation/ui/tests/home/HelpButtonTest.java
@@ -1,5 +1,6 @@
 package com.generic.functional.automation.ui.tests.home;
 
+import com.aventstack.extentreports.Status;
 import com.generic.functional.automation.ui.tests.common.TestConfig;
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
@@ -80,4 +81,27 @@ public class HelpButtonTest extends TestConfig {
 
 
     }
+    @Test (groups = {"smokeTest"})
+    public void testHelpButtonShipmentsClick() throws Exception {
+        test = extent.createTest("Help Button Shipments Click");
+        login.doLogin(test);
+        Thread.sleep(5 * 1000);
+        test.log(Status.INFO, "Looking for help button");
+        WebElement helpButton = driver.findElement(By.cssSelector(".explore-quiries-inner"));
+        helpButton.click();
+        test.log(Status.INFO, "Help Button Clicked");
+        driver.findElement(By.id("mainArc-0b981a1b-32dc-43b1-b257-70c8c5a6cc6d")).click();
+        Thread.sleep(5 * 1000);
+        test.log(Status.INFO, "Shipments cluster found");
+        driver.findElement(By.id("mainArc-d5f6e57b-d157-4ec5-bbc4-9fd35cc68775")).click();
+        Thread.sleep(5 * 1000);
+        test.log(Status.INFO, "Service cluster found");
+        driver.findElement(By.id("mainArc-f18c2bb1-ad7d-4583-8b64-7ef98ba647cc")).click();
+        Thread.sleep(5 * 1000);
+        driver.findElement(By.xpath("//*[text()=' Total Records']"));
+
+        test.createNode("Total Records found means table loaded");
+
+    }
+
 }
